Match DrawPrevNextCursor
The page count is an int, the page step sits in the loop's increment, and the last page ends the draw. Its positions template now comes from menu_manual.

+++ include/menu_manual.hpp
@@ -50,7 +50,7 @@ extern ClsMes *ManualMsg;
  * @address 0x00233520
  * @size 0x38
  */
-s8 GetNowManualPartTgaNum();
+int GetNowManualPartTgaNum();
 
 /**
  * Reports whether the party has the item that unlocks the manual.
